Network.AWS.DMS.CreateEndpoint

Description

Creates an endpoint using the provided settings.

Creating a Request

createEndpoint

Arguments

:: Text
ceEndpointIdentifier

-> ReplicationEndpointTypeValue
ceEndpointType

-> Text
ceEngineName

-> CreateEndpoint

Creates a value of CreateEndpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ceServerName - The name of the server where the endpoint database resides.

ceCertificateARN - The Amazon Resource Number (ARN) for the certificate.

ceExtraConnectionAttributes - Additional attributes associated with the connection.

ceUsername - The user name to be used to login to the endpoint database.

ceKMSKeyId - The KMS key identifier that will be used to encrypt the connection parameters. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

ceMongoDBSettings - Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the Configuration Properties When Using MongoDB as a Source for AWS Database Migration Service section at Using Amazon S3 as a Target for AWS Database Migration Service .

ceSSLMode - The SSL mode to use for the SSL connection. SSL mode can be one of four values: none, require, verify-ca, verify-full. The default value is none.

cePassword - The password to be used to login to the endpoint database.

ceDatabaseName - The name of the endpoint database.

ceS3Settings - Settings in JSON format for the target S3 endpoint. For more information about the available settings, see the Extra Connection Attributes section at Using Amazon S3 as a Target for AWS Database Migration Service .

ceDynamoDBSettings - Settings in JSON format for the target Amazon DynamoDB endpoint. For more information about the available settings, see the Using Object Mapping to Migrate Data to DynamoDB section at Using an Amazon DynamoDB Database as a Target for AWS Database Migration Service .

ceTags - Tags to be added to the endpoint.

cePort - The port used by the endpoint database.

ceEndpointIdentifier - The database endpoint identifier. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.

ceEndpointType - The type of endpoint.

ceEngineName - The type of engine for the endpoint. Valid values, depending on the EndPointType, include MYSQL, ORACLE, POSTGRES, MARIADB, AURORA, REDSHIFT, S3, SYBASE, DYNAMODB, MONGODB, and SQLSERVER.

data CreateEndpoint

See: createEndpoint smart constructor.

Request Lenses

ceServerName :: Lens' CreateEndpoint (Maybe Text)

The name of the server where the endpoint database resides.

ceCertificateARN :: Lens' CreateEndpoint (Maybe Text)

The Amazon Resource Number (ARN) for the certificate.

ceExtraConnectionAttributes :: Lens' CreateEndpoint (Maybe Text)

Additional attributes associated with the connection.

ceUsername :: Lens' CreateEndpoint (Maybe Text)

The user name to be used to login to the endpoint database.

ceKMSKeyId :: Lens' CreateEndpoint (Maybe Text)

The KMS key identifier that will be used to encrypt the connection parameters. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

ceMongoDBSettings :: Lens' CreateEndpoint (Maybe MongoDBSettings)

Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the Configuration Properties When Using MongoDB as a Source for AWS Database Migration Service section at Using Amazon S3 as a Target for AWS Database Migration Service .

ceSSLMode :: Lens' CreateEndpoint (Maybe DmsSSLModeValue)

The SSL mode to use for the SSL connection. SSL mode can be one of four values: none, require, verify-ca, verify-full. The default value is none.

cePassword :: Lens' CreateEndpoint (Maybe Text)

The password to be used to login to the endpoint database.

ceDatabaseName :: Lens' CreateEndpoint (Maybe Text)

The name of the endpoint database.

ceS3Settings :: Lens' CreateEndpoint (Maybe S3Settings)

Settings in JSON format for the target S3 endpoint. For more information about the available settings, see the Extra Connection Attributes section at Using Amazon S3 as a Target for AWS Database Migration Service .

ceDynamoDBSettings :: Lens' CreateEndpoint (Maybe DynamoDBSettings)

Settings in JSON format for the target Amazon DynamoDB endpoint. For more information about the available settings, see the Using Object Mapping to Migrate Data to DynamoDB section at Using an Amazon DynamoDB Database as a Target for AWS Database Migration Service .

ceTags :: Lens' CreateEndpoint [Tag]

Tags to be added to the endpoint.

cePort :: Lens' CreateEndpoint (Maybe Int)

The port used by the endpoint database.

ceEndpointIdentifier :: Lens' CreateEndpoint Text

The database endpoint identifier. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.

ceEndpointType :: Lens' CreateEndpoint ReplicationEndpointTypeValue

The type of endpoint.

ceEngineName :: Lens' CreateEndpoint Text

The type of engine for the endpoint. Valid values, depending on the EndPointType, include MYSQL, ORACLE, POSTGRES, MARIADB, AURORA, REDSHIFT, S3, SYBASE, DYNAMODB, MONGODB, and SQLSERVER.

Destructuring the Response

createEndpointResponse

Arguments

:: Int
cersResponseStatus

-> CreateEndpointResponse

Creates a value of CreateEndpointResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cersEndpoint - The endpoint that was created.

cersResponseStatus - -- | The response status code.

data CreateEndpointResponse

See: createEndpointResponse smart constructor.

Response Lenses

cersEndpoint :: Lens' CreateEndpointResponse (Maybe Endpoint)

The endpoint that was created.

cersResponseStatus :: Lens' CreateEndpointResponse Int

- | The response status code.