Network.AWS.DMS.CreateReplicationInstance

Description

Creates the replication instance using the specified parameters.

Creating a Request

createReplicationInstance

Arguments

:: Text
criReplicationInstanceIdentifier

-> Text
criReplicationInstanceClass

-> CreateReplicationInstance

Creates a value of CreateReplicationInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

criEngineVersion - The engine version number of the replication instance.

criPubliclyAccessible - Specifies the accessibility options for the replication instance. A value of true represents an instance with a public IP address. A value of false represents an instance with a private IP address. The default value is true .

criAutoMinorVersionUpgrade - Indicates that minor engine upgrades will be applied automatically to the replication instance during the maintenance window. Default: true

criReplicationSubnetGroupIdentifier - A subnet group to associate with the replication instance.

criPreferredMaintenanceWindow - The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC). Format: ddd:hh24:mi-ddd:hh24:mi Default: A 30-minute window selected at random from an 8-hour block of time per region, occurring on a random day of the week. Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun Constraints: Minimum 30-minute window.

criKMSKeyId - The KMS key identifier that will be used to encrypt the content on the replication instance. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

criAvailabilityZone - The EC2 Availability Zone that the replication instance will be created in. Default: A random, system-chosen Availability Zone in the endpoint's region. Example: us-east-1d

criVPCSecurityGroupIds - Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

criMultiAZ - Specifies if the replication instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the Multi-AZ parameter is set to true .

criAllocatedStorage - The amount of storage (in gigabytes) to be initially allocated for the replication instance.

criTags - Tags to be associated with the replication instance.

criReplicationInstanceIdentifier - The replication instance identifier. This parameter is stored as a lowercase string. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. Example: myrepinstance

criReplicationInstanceClass - The compute and memory capacity of the replication instance as specified by the replication instance class. Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

data CreateReplicationInstance

See: createReplicationInstance smart constructor.

Request Lenses

criEngineVersion :: Lens' CreateReplicationInstance (Maybe Text)

The engine version number of the replication instance.

criPubliclyAccessible :: Lens' CreateReplicationInstance (Maybe Bool)

Specifies the accessibility options for the replication instance. A value of true represents an instance with a public IP address. A value of false represents an instance with a private IP address. The default value is true .

criAutoMinorVersionUpgrade :: Lens' CreateReplicationInstance (Maybe Bool)

Indicates that minor engine upgrades will be applied automatically to the replication instance during the maintenance window. Default: true

criReplicationSubnetGroupIdentifier :: Lens' CreateReplicationInstance (Maybe Text)

A subnet group to associate with the replication instance.

criPreferredMaintenanceWindow :: Lens' CreateReplicationInstance (Maybe Text)

The weekly time range during which system maintenance can occur, in Universal Coordinated Time (UTC). Format: ddd:hh24:mi-ddd:hh24:mi Default: A 30-minute window selected at random from an 8-hour block of time per region, occurring on a random day of the week. Valid Days: Mon, Tue, Wed, Thu, Fri, Sat, Sun Constraints: Minimum 30-minute window.

criKMSKeyId :: Lens' CreateReplicationInstance (Maybe Text)

The KMS key identifier that will be used to encrypt the content on the replication instance. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

criAvailabilityZone :: Lens' CreateReplicationInstance (Maybe Text)

The EC2 Availability Zone that the replication instance will be created in. Default: A random, system-chosen Availability Zone in the endpoint's region. Example: us-east-1d

criVPCSecurityGroupIds :: Lens' CreateReplicationInstance [Text]

Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

criMultiAZ :: Lens' CreateReplicationInstance (Maybe Bool)

Specifies if the replication instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the Multi-AZ parameter is set to true .

criAllocatedStorage :: Lens' CreateReplicationInstance (Maybe Int)

The amount of storage (in gigabytes) to be initially allocated for the replication instance.

criTags :: Lens' CreateReplicationInstance [Tag]

Tags to be associated with the replication instance.

criReplicationInstanceIdentifier :: Lens' CreateReplicationInstance Text

The replication instance identifier. This parameter is stored as a lowercase string. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. Example: myrepinstance

criReplicationInstanceClass :: Lens' CreateReplicationInstance Text

The compute and memory capacity of the replication instance as specified by the replication instance class. Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

Destructuring the Response

createReplicationInstanceResponse

Arguments

:: Int
crirsResponseStatus

-> CreateReplicationInstanceResponse

Creates a value of CreateReplicationInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crirsReplicationInstance - The replication instance that was created.

crirsResponseStatus - -- | The response status code.

data CreateReplicationInstanceResponse

See: createReplicationInstanceResponse smart constructor.

Response Lenses

crirsReplicationInstance :: Lens' CreateReplicationInstanceResponse (Maybe ReplicationInstance)

The replication instance that was created.

crirsResponseStatus :: Lens' CreateReplicationInstanceResponse Int

- | The response status code.