Network.AWS.DMS.CreateReplicationSubnetGroup

Description

Creates a replication subnet group given a list of the subnet IDs in a VPC.

Creating a Request

createReplicationSubnetGroup

Arguments

:: Text
crsgReplicationSubnetGroupIdentifier

-> Text
crsgReplicationSubnetGroupDescription

-> CreateReplicationSubnetGroup

Creates a value of CreateReplicationSubnetGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crsgTags - The tag to be assigned to the subnet group.

crsgReplicationSubnetGroupIdentifier - The name for the replication subnet group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters, periods, spaces, underscores, or hyphens. Must not be "default". Example: mySubnetgroup

crsgReplicationSubnetGroupDescription - The description for the subnet group.

crsgSubnetIds - The EC2 subnet IDs for the subnet group.

data CreateReplicationSubnetGroup

See: createReplicationSubnetGroup smart constructor.

Request Lenses

crsgTags :: Lens' CreateReplicationSubnetGroup [Tag]

The tag to be assigned to the subnet group.

crsgReplicationSubnetGroupIdentifier :: Lens' CreateReplicationSubnetGroup Text

The name for the replication subnet group. This value is stored as a lowercase string. Constraints: Must contain no more than 255 alphanumeric characters, periods, spaces, underscores, or hyphens. Must not be "default". Example: mySubnetgroup

crsgReplicationSubnetGroupDescription :: Lens' CreateReplicationSubnetGroup Text

The description for the subnet group.

crsgSubnetIds :: Lens' CreateReplicationSubnetGroup [Text]

The EC2 subnet IDs for the subnet group.

Destructuring the Response

createReplicationSubnetGroupResponse

Arguments

:: Int
crsgrsResponseStatus

-> CreateReplicationSubnetGroupResponse

Creates a value of CreateReplicationSubnetGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crsgrsReplicationSubnetGroup - The replication subnet group that was created.

crsgrsResponseStatus - -- | The response status code.

data CreateReplicationSubnetGroupResponse

See: createReplicationSubnetGroupResponse smart constructor.

Response Lenses

crsgrsReplicationSubnetGroup :: Lens' CreateReplicationSubnetGroupResponse (Maybe ReplicationSubnetGroup)

The replication subnet group that was created.

crsgrsResponseStatus :: Lens' CreateReplicationSubnetGroupResponse Int

- | The response status code.