Network.AWS.DMS.CreateReplicationTask

Description

Creates a replication task using the specified parameters.

Creating a Request

createReplicationTask

Arguments

:: Text
crtReplicationTaskIdentifier

-> Text
crtSourceEndpointARN

-> Text
crtTargetEndpointARN

-> Text
crtReplicationInstanceARN

-> MigrationTypeValue
crtMigrationType

-> Text
crtTableMappings

-> CreateReplicationTask

Creates a value of CreateReplicationTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crtReplicationTaskSettings - Settings for the task, such as target metadata settings. For a complete list of task settings, see Task Settings for AWS Database Migration Service Tasks .

crtTags - Tags to be added to the replication instance.

crtCdcStartTime - The start time for the Change Data Capture (CDC) operation.

crtReplicationTaskIdentifier - The replication task identifier. Constraints: * Must contain from 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

crtSourceEndpointARN - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

crtTargetEndpointARN - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

crtReplicationInstanceARN - The Amazon Resource Name (ARN) of the replication instance.

crtMigrationType - The migration type.

crtTableMappings - When using the AWS CLI or boto3, provide the path of the JSON file that contains the table mappings. Precede the path with "file:/". When working with the DMS API, provide the JSON as the parameter value. For example, --table-mappings file:/mappingfile.json

data CreateReplicationTask

See: createReplicationTask smart constructor.

Request Lenses

crtReplicationTaskSettings :: Lens' CreateReplicationTask (Maybe Text)

Settings for the task, such as target metadata settings. For a complete list of task settings, see Task Settings for AWS Database Migration Service Tasks .

crtTags :: Lens' CreateReplicationTask [Tag]

Tags to be added to the replication instance.

crtCdcStartTime :: Lens' CreateReplicationTask (Maybe UTCTime)

The start time for the Change Data Capture (CDC) operation.

crtReplicationTaskIdentifier :: Lens' CreateReplicationTask Text

The replication task identifier. Constraints: * Must contain from 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

crtSourceEndpointARN :: Lens' CreateReplicationTask Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

crtTargetEndpointARN :: Lens' CreateReplicationTask Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

crtReplicationInstanceARN :: Lens' CreateReplicationTask Text

The Amazon Resource Name (ARN) of the replication instance.

crtMigrationType :: Lens' CreateReplicationTask MigrationTypeValue

The migration type.

crtTableMappings :: Lens' CreateReplicationTask Text

When using the AWS CLI or boto3, provide the path of the JSON file that contains the table mappings. Precede the path with "file:/". When working with the DMS API, provide the JSON as the parameter value. For example, --table-mappings file:/mappingfile.json

Destructuring the Response

createReplicationTaskResponse

Arguments

:: Int
crtrsResponseStatus

-> CreateReplicationTaskResponse

Creates a value of CreateReplicationTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crtrsReplicationTask - The replication task that was created.

crtrsResponseStatus - -- | The response status code.

data CreateReplicationTaskResponse

See: createReplicationTaskResponse smart constructor.

Response Lenses

crtrsReplicationTask :: Lens' CreateReplicationTaskResponse (Maybe ReplicationTask)

The replication task that was created.

crtrsResponseStatus :: Lens' CreateReplicationTaskResponse Int

- | The response status code.