Network.AWS.DMS.DeleteCertificate

Description

Deletes the specified certificate.

Creating a Request

deleteCertificate

Arguments

:: Text
dcCertificateARN

-> DeleteCertificate

Creates a value of DeleteCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcCertificateARN - The Amazon Resource Name (ARN) of the deleted certificate.

data DeleteCertificate

See: deleteCertificate smart constructor.

Request Lenses

dcCertificateARN :: Lens' DeleteCertificate Text

The Amazon Resource Name (ARN) of the deleted certificate.

Destructuring the Response

deleteCertificateResponse

Arguments

:: Int
dccrsResponseStatus

-> DeleteCertificateResponse

Creates a value of DeleteCertificateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dccrsCertificate - The Secure Sockets Layer (SSL) certificate.

dccrsResponseStatus - -- | The response status code.

data DeleteCertificateResponse

See: deleteCertificateResponse smart constructor.

Response Lenses

dccrsCertificate :: Lens' DeleteCertificateResponse (Maybe Certificate)

The Secure Sockets Layer (SSL) certificate.

dccrsResponseStatus :: Lens' DeleteCertificateResponse Int

- | The response status code.