Network.AWS.DMS.DescribeConnections

Description

Describes the status of the connections that have been made between the replication instance and an endpoint. Connections are created when you test an endpoint.

Creating a Request

describeConnections :: DescribeConnections

Creates a value of DescribeConnections with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcFilters - The filters applied to the connection. Valid filter names: endpoint-arn | replication-instance-arn

dcMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

data DescribeConnections

See: describeConnections smart constructor.

Request Lenses

dcFilters :: Lens' DescribeConnections [Filter]

The filters applied to the connection. Valid filter names: endpoint-arn | replication-instance-arn

dcMarker :: Lens' DescribeConnections (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dcMaxRecords :: Lens' DescribeConnections (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

Destructuring the Response

describeConnectionsResponse

Arguments

:: Int
drsResponseStatus

-> DescribeConnectionsResponse

Creates a value of DescribeConnectionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsConnections - A description of the connections.

drsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

drsResponseStatus - -- | The response status code.

data DescribeConnectionsResponse

See: describeConnectionsResponse smart constructor.

Response Lenses

drsConnections :: Lens' DescribeConnectionsResponse [Connection]

A description of the connections.

drsMarker :: Lens' DescribeConnectionsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

drsResponseStatus :: Lens' DescribeConnectionsResponse Int

- | The response status code.