Network.AWS.DMS.DescribeEndpoints

Description

Returns information about the endpoints for your account in the current region.

Creating a Request

describeEndpoints :: DescribeEndpoints

Creates a value of DescribeEndpoints with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desFilters - Filters applied to the describe action. Valid filter names: endpoint-arn | endpoint-type | endpoint-id | engine-name

desMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

desMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

data DescribeEndpoints

See: describeEndpoints smart constructor.

Request Lenses

desFilters :: Lens' DescribeEndpoints [Filter]

Filters applied to the describe action. Valid filter names: endpoint-arn | endpoint-type | endpoint-id | engine-name

desMarker :: Lens' DescribeEndpoints (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

desMaxRecords :: Lens' DescribeEndpoints (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

Destructuring the Response

describeEndpointsResponse

Arguments

:: Int
dersResponseStatus

-> DescribeEndpointsResponse

Creates a value of DescribeEndpointsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dersMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dersEndpoints - Endpoint description.

dersResponseStatus - -- | The response status code.

data DescribeEndpointsResponse

See: describeEndpointsResponse smart constructor.

Response Lenses

dersMarker :: Lens' DescribeEndpointsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dersEndpoints :: Lens' DescribeEndpointsResponse [Endpoint]

Endpoint description.

dersResponseStatus :: Lens' DescribeEndpointsResponse Int

- | The response status code.