Network.AWS.DMS.DescribeEventSubscriptions

Description

Lists all the event subscriptions for a customer account. The description of a subscription includes SubscriptionName , SNSTopicARN , CustomerID , SourceType , SourceID , CreationTime , and Status .

If you specify SubscriptionName , this action lists the description for that subscription.

Creating a Request

describeEventSubscriptions :: DescribeEventSubscriptions

Creates a value of DescribeEventSubscriptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dessSubscriptionName - The name of the AWS DMS event subscription to be described.

dessFilters - Filters applied to the action.

dessMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dessMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

data DescribeEventSubscriptions

See: describeEventSubscriptions smart constructor.

Request Lenses

dessSubscriptionName :: Lens' DescribeEventSubscriptions (Maybe Text)

The name of the AWS DMS event subscription to be described.

dessFilters :: Lens' DescribeEventSubscriptions [Filter]

Filters applied to the action.

dessMarker :: Lens' DescribeEventSubscriptions (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dessMaxRecords :: Lens' DescribeEventSubscriptions (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

Destructuring the Response

describeEventSubscriptionsResponse

Arguments

:: Int
desrsResponseStatus

-> DescribeEventSubscriptionsResponse

Creates a value of DescribeEventSubscriptionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desrsEventSubscriptionsList - A list of event subscriptions.

desrsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

desrsResponseStatus - -- | The response status code.

data DescribeEventSubscriptionsResponse

See: describeEventSubscriptionsResponse smart constructor.

Response Lenses

desrsEventSubscriptionsList :: Lens' DescribeEventSubscriptionsResponse [EventSubscription]

A list of event subscriptions.

desrsMarker :: Lens' DescribeEventSubscriptionsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

desrsResponseStatus :: Lens' DescribeEventSubscriptionsResponse Int

- | The response status code.