Network.AWS.DMS.DescribeOrderableReplicationInstances

Description

Returns information about the replication instance types that can be created in the specified region.

Creating a Request

describeOrderableReplicationInstances :: DescribeOrderableReplicationInstances

Creates a value of DescribeOrderableReplicationInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

doriMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

doriMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

data DescribeOrderableReplicationInstances

See: describeOrderableReplicationInstances smart constructor.

Request Lenses

doriMarker :: Lens' DescribeOrderableReplicationInstances (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

doriMaxRecords :: Lens' DescribeOrderableReplicationInstances (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

Destructuring the Response

describeOrderableReplicationInstancesResponse

Arguments

:: Int
dorirsResponseStatus

-> DescribeOrderableReplicationInstancesResponse

Creates a value of DescribeOrderableReplicationInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dorirsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dorirsOrderableReplicationInstances - The order-able replication instances available.

dorirsResponseStatus - -- | The response status code.

data DescribeOrderableReplicationInstancesResponse

See: describeOrderableReplicationInstancesResponse smart constructor.

Response Lenses

dorirsMarker :: Lens' DescribeOrderableReplicationInstancesResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dorirsOrderableReplicationInstances :: Lens' DescribeOrderableReplicationInstancesResponse [OrderableReplicationInstance]

The order-able replication instances available.

dorirsResponseStatus :: Lens' DescribeOrderableReplicationInstancesResponse Int

- | The response status code.