Network.AWS.DMS.DescribeSchemas

Description

Returns information about the schema for the specified endpoint.

Creating a Request

describeSchemas

Arguments

:: Text
dsEndpointARN

-> DescribeSchemas

Creates a value of DescribeSchemas with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dsMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

dsEndpointARN - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

data DescribeSchemas

See: describeSchemas smart constructor.

Request Lenses

dsMarker :: Lens' DescribeSchemas (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dsMaxRecords :: Lens' DescribeSchemas (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

dsEndpointARN :: Lens' DescribeSchemas Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

Destructuring the Response

describeSchemasResponse

Arguments

:: Int
dsrsResponseStatus

-> DescribeSchemasResponse

Creates a value of DescribeSchemasResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsrsSchemas - The described schema.

dsrsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dsrsResponseStatus - -- | The response status code.

data DescribeSchemasResponse

See: describeSchemasResponse smart constructor.

Response Lenses

dsrsSchemas :: Lens' DescribeSchemasResponse [Text]

The described schema.

dsrsMarker :: Lens' DescribeSchemasResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dsrsResponseStatus :: Lens' DescribeSchemasResponse Int

- | The response status code.