Network.AWS.DMS.DescribeTableStatistics

Description

Returns table statistics on the database migration task, including table name, rows inserted, rows updated, and rows deleted.

Creating a Request

describeTableStatistics

Arguments

:: Text
dtsReplicationTaskARN

-> DescribeTableStatistics

Creates a value of DescribeTableStatistics with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dtsMaxRecords - The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

dtsReplicationTaskARN - The Amazon Resource Name (ARN) of the replication task.

data DescribeTableStatistics

See: describeTableStatistics smart constructor.

Request Lenses

dtsMarker :: Lens' DescribeTableStatistics (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dtsMaxRecords :: Lens' DescribeTableStatistics (Maybe Int)

The maximum number of records to include in the response. If more records exist than the specified MaxRecords value, a pagination token called a marker is included in the response so that the remaining results can be retrieved. Default: 100 Constraints: Minimum 20, maximum 100.

dtsReplicationTaskARN :: Lens' DescribeTableStatistics Text

The Amazon Resource Name (ARN) of the replication task.

Destructuring the Response

describeTableStatisticsResponse

Arguments

:: Int
dtsrsResponseStatus

-> DescribeTableStatisticsResponse

Creates a value of DescribeTableStatisticsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtsrsReplicationTaskARN - The Amazon Resource Name (ARN) of the replication task.

dtsrsMarker - An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dtsrsTableStatistics - The table statistics.

dtsrsResponseStatus - -- | The response status code.

data DescribeTableStatisticsResponse

See: describeTableStatisticsResponse smart constructor.

Response Lenses

dtsrsReplicationTaskARN :: Lens' DescribeTableStatisticsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the replication task.

dtsrsMarker :: Lens' DescribeTableStatisticsResponse (Maybe Text)

An optional pagination token provided by a previous request. If this parameter is specified, the response includes only records beyond the marker, up to the value specified by MaxRecords .

dtsrsTableStatistics :: Lens' DescribeTableStatisticsResponse [TableStatistics]

The table statistics.

dtsrsResponseStatus :: Lens' DescribeTableStatisticsResponse Int

- | The response status code.