Network.AWS.DMS.ImportCertificate

Description

Uploads the specified certificate.

Creating a Request

importCertificate

Arguments

:: Text
icCertificateIdentifier

-> ImportCertificate

Creates a value of ImportCertificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

icCertificatePem - The contents of the .pem X.509 certificate file for the certificate.

icCertificateWallet - The location of the imported Oracle Wallet certificate for use with SSL.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

icTags - The tags associated with the certificate.

icCertificateIdentifier - The customer-assigned name of the certificate. Valid characters are A-z and 0-9.

data ImportCertificate

See: importCertificate smart constructor.

Request Lenses

icCertificatePem :: Lens' ImportCertificate (Maybe Text)

The contents of the .pem X.509 certificate file for the certificate.

icCertificateWallet :: Lens' ImportCertificate (Maybe ByteString)

The location of the imported Oracle Wallet certificate for use with SSL.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

icTags :: Lens' ImportCertificate [Tag]

The tags associated with the certificate.

icCertificateIdentifier :: Lens' ImportCertificate Text

The customer-assigned name of the certificate. Valid characters are A-z and 0-9.

Destructuring the Response

importCertificateResponse

Arguments

:: Int
icrsResponseStatus

-> ImportCertificateResponse

Creates a value of ImportCertificateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

icrsCertificate - The certificate to be uploaded.

icrsResponseStatus - -- | The response status code.

data ImportCertificateResponse

See: importCertificateResponse smart constructor.

Response Lenses

icrsCertificate :: Lens' ImportCertificateResponse (Maybe Certificate)

The certificate to be uploaded.

icrsResponseStatus :: Lens' ImportCertificateResponse Int

- | The response status code.