Network.AWS.DMS.ModifyEndpoint

Description

Modifies the specified endpoint.

Creating a Request

modifyEndpoint

Arguments

:: Text
meEndpointARN

-> ModifyEndpoint

Creates a value of ModifyEndpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

meServerName - The name of the server where the endpoint database resides.

meCertificateARN - The Amazon Resource Name (ARN) of the certificate used for SSL connection.

meExtraConnectionAttributes - Additional attributes associated with the connection.

meEndpointType - The type of endpoint.

meUsername - The user name to be used to login to the endpoint database.

meEngineName - The type of engine for the endpoint. Valid values, depending on the EndPointType, include MYSQL, ORACLE, POSTGRES, MARIADB, AURORA, REDSHIFT, S3, DYNAMODB, MONGODB, SYBASE, and SQLSERVER.

meMongoDBSettings - Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the Configuration Properties When Using MongoDB as a Source for AWS Database Migration Service section at Using Amazon S3 as a Target for AWS Database Migration Service .

meSSLMode - The SSL mode to be used. SSL mode can be one of four values: none, require, verify-ca, verify-full. The default value is none.

mePassword - The password to be used to login to the endpoint database.

meDatabaseName - The name of the endpoint database.

meS3Settings - Settings in JSON format for the target S3 endpoint. For more information about the available settings, see the Extra Connection Attributes section at Using Amazon S3 as a Target for AWS Database Migration Service .

meEndpointIdentifier - The database endpoint identifier. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.

meDynamoDBSettings - Settings in JSON format for the target Amazon DynamoDB endpoint. For more information about the available settings, see the Using Object Mapping to Migrate Data to DynamoDB section at Using an Amazon DynamoDB Database as a Target for AWS Database Migration Service .

mePort - The port used by the endpoint database.

meEndpointARN - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

data ModifyEndpoint

See: modifyEndpoint smart constructor.

Request Lenses

meServerName :: Lens' ModifyEndpoint (Maybe Text)

The name of the server where the endpoint database resides.

meCertificateARN :: Lens' ModifyEndpoint (Maybe Text)

The Amazon Resource Name (ARN) of the certificate used for SSL connection.

meExtraConnectionAttributes :: Lens' ModifyEndpoint (Maybe Text)

Additional attributes associated with the connection.

meEndpointType :: Lens' ModifyEndpoint (Maybe ReplicationEndpointTypeValue)

The type of endpoint.

meUsername :: Lens' ModifyEndpoint (Maybe Text)

The user name to be used to login to the endpoint database.

meEngineName :: Lens' ModifyEndpoint (Maybe Text)

The type of engine for the endpoint. Valid values, depending on the EndPointType, include MYSQL, ORACLE, POSTGRES, MARIADB, AURORA, REDSHIFT, S3, DYNAMODB, MONGODB, SYBASE, and SQLSERVER.

meMongoDBSettings :: Lens' ModifyEndpoint (Maybe MongoDBSettings)

Settings in JSON format for the source MongoDB endpoint. For more information about the available settings, see the Configuration Properties When Using MongoDB as a Source for AWS Database Migration Service section at Using Amazon S3 as a Target for AWS Database Migration Service .

meSSLMode :: Lens' ModifyEndpoint (Maybe DmsSSLModeValue)

The SSL mode to be used. SSL mode can be one of four values: none, require, verify-ca, verify-full. The default value is none.

mePassword :: Lens' ModifyEndpoint (Maybe Text)

The password to be used to login to the endpoint database.

meDatabaseName :: Lens' ModifyEndpoint (Maybe Text)

The name of the endpoint database.

meS3Settings :: Lens' ModifyEndpoint (Maybe S3Settings)

Settings in JSON format for the target S3 endpoint. For more information about the available settings, see the Extra Connection Attributes section at Using Amazon S3 as a Target for AWS Database Migration Service .

meEndpointIdentifier :: Lens' ModifyEndpoint (Maybe Text)

The database endpoint identifier. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.

meDynamoDBSettings :: Lens' ModifyEndpoint (Maybe DynamoDBSettings)

Settings in JSON format for the target Amazon DynamoDB endpoint. For more information about the available settings, see the Using Object Mapping to Migrate Data to DynamoDB section at Using an Amazon DynamoDB Database as a Target for AWS Database Migration Service .

mePort :: Lens' ModifyEndpoint (Maybe Int)

The port used by the endpoint database.

meEndpointARN :: Lens' ModifyEndpoint Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

Destructuring the Response

modifyEndpointResponse

Arguments

:: Int
mersResponseStatus

-> ModifyEndpointResponse

Creates a value of ModifyEndpointResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mersEndpoint - The modified endpoint.

mersResponseStatus - -- | The response status code.

data ModifyEndpointResponse

See: modifyEndpointResponse smart constructor.

Response Lenses

mersEndpoint :: Lens' ModifyEndpointResponse (Maybe Endpoint)

The modified endpoint.

mersResponseStatus :: Lens' ModifyEndpointResponse Int

- | The response status code.