Network.AWS.DMS.ModifyReplicationInstance

Description

Modifies the replication instance to apply new settings. You can change one or more parameters by specifying these parameters and the new values in the request.

Some settings are applied during the maintenance window.

Creating a Request

modifyReplicationInstance

Arguments

:: Text
mriReplicationInstanceARN

-> ModifyReplicationInstance

Creates a value of ModifyReplicationInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mriEngineVersion - The engine version number of the replication instance.

mriAutoMinorVersionUpgrade - Indicates that minor version upgrades will be applied automatically to the replication instance during the maintenance window. Changing this parameter does not result in an outage except in the following case and the change is asynchronously applied as soon as possible. An outage will result if this parameter is set to true during the maintenance window, and a newer minor version is available, and AWS DMS has enabled auto patching for that engine version.

mriAllowMajorVersionUpgrade - Indicates that major version upgrades are allowed. Changing this parameter does not result in an outage and the change is asynchronously applied as soon as possible. Constraints: This parameter must be set to true when specifying a value for the EngineVersion parameter that is a different major version than the replication instance's current version.

mriPreferredMaintenanceWindow - The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter does not result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied. Default: Uses existing setting Format: ddd:hh24:mi-ddd:hh24:mi Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun Constraints: Must be at least 30 minutes

mriVPCSecurityGroupIds - Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

mriMultiAZ - Specifies if the replication instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the Multi-AZ parameter is set to true .

mriAllocatedStorage - The amount of storage (in gigabytes) to be allocated for the replication instance.

mriApplyImmediately - Indicates whether the changes should be applied immediately or during the next maintenance window.

mriReplicationInstanceClass - The compute and memory capacity of the replication instance. Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

mriReplicationInstanceIdentifier - The replication instance identifier. This parameter is stored as a lowercase string.

mriReplicationInstanceARN - The Amazon Resource Name (ARN) of the replication instance.

data ModifyReplicationInstance

See: modifyReplicationInstance smart constructor.

Request Lenses

mriEngineVersion :: Lens' ModifyReplicationInstance (Maybe Text)

The engine version number of the replication instance.

mriAutoMinorVersionUpgrade :: Lens' ModifyReplicationInstance (Maybe Bool)

Indicates that minor version upgrades will be applied automatically to the replication instance during the maintenance window. Changing this parameter does not result in an outage except in the following case and the change is asynchronously applied as soon as possible. An outage will result if this parameter is set to true during the maintenance window, and a newer minor version is available, and AWS DMS has enabled auto patching for that engine version.

mriAllowMajorVersionUpgrade :: Lens' ModifyReplicationInstance (Maybe Bool)

Indicates that major version upgrades are allowed. Changing this parameter does not result in an outage and the change is asynchronously applied as soon as possible. Constraints: This parameter must be set to true when specifying a value for the EngineVersion parameter that is a different major version than the replication instance's current version.

mriPreferredMaintenanceWindow :: Lens' ModifyReplicationInstance (Maybe Text)

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter does not result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied. Default: Uses existing setting Format: ddd:hh24:mi-ddd:hh24:mi Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun Constraints: Must be at least 30 minutes

mriVPCSecurityGroupIds :: Lens' ModifyReplicationInstance [Text]

Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.

mriMultiAZ :: Lens' ModifyReplicationInstance (Maybe Bool)

Specifies if the replication instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the Multi-AZ parameter is set to true .

mriAllocatedStorage :: Lens' ModifyReplicationInstance (Maybe Int)

The amount of storage (in gigabytes) to be allocated for the replication instance.

mriApplyImmediately :: Lens' ModifyReplicationInstance (Maybe Bool)

Indicates whether the changes should be applied immediately or during the next maintenance window.

mriReplicationInstanceClass :: Lens' ModifyReplicationInstance (Maybe Text)

The compute and memory capacity of the replication instance. Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

mriReplicationInstanceIdentifier :: Lens' ModifyReplicationInstance (Maybe Text)

The replication instance identifier. This parameter is stored as a lowercase string.

mriReplicationInstanceARN :: Lens' ModifyReplicationInstance Text

The Amazon Resource Name (ARN) of the replication instance.

Destructuring the Response

modifyReplicationInstanceResponse

Arguments

:: Int
mrirsResponseStatus

-> ModifyReplicationInstanceResponse

Creates a value of ModifyReplicationInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mrirsReplicationInstance - The modified replication instance.

mrirsResponseStatus - -- | The response status code.

data ModifyReplicationInstanceResponse

See: modifyReplicationInstanceResponse smart constructor.

Response Lenses

mrirsReplicationInstance :: Lens' ModifyReplicationInstanceResponse (Maybe ReplicationInstance)

The modified replication instance.

mrirsResponseStatus :: Lens' ModifyReplicationInstanceResponse Int

- | The response status code.