Network.AWS.DMS.ModifyReplicationSubnetGroup

Description

Modifies the settings for the specified replication subnet group.

Creating a Request

modifyReplicationSubnetGroup

Arguments

:: Text
mrsgReplicationSubnetGroupIdentifier

-> ModifyReplicationSubnetGroup

Creates a value of ModifyReplicationSubnetGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mrsgReplicationSubnetGroupDescription - The description of the replication instance subnet group.

mrsgReplicationSubnetGroupIdentifier - The name of the replication instance subnet group.

mrsgSubnetIds - A list of subnet IDs.

data ModifyReplicationSubnetGroup

See: modifyReplicationSubnetGroup smart constructor.

Request Lenses

mrsgReplicationSubnetGroupDescription :: Lens' ModifyReplicationSubnetGroup (Maybe Text)

The description of the replication instance subnet group.

mrsgReplicationSubnetGroupIdentifier :: Lens' ModifyReplicationSubnetGroup Text

The name of the replication instance subnet group.

mrsgSubnetIds :: Lens' ModifyReplicationSubnetGroup [Text]

A list of subnet IDs.

Destructuring the Response

modifyReplicationSubnetGroupResponse

Arguments

:: Int
mrsgrsResponseStatus

-> ModifyReplicationSubnetGroupResponse

Creates a value of ModifyReplicationSubnetGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mrsgrsReplicationSubnetGroup - The modified replication subnet group.

mrsgrsResponseStatus - -- | The response status code.

data ModifyReplicationSubnetGroupResponse

See: modifyReplicationSubnetGroupResponse smart constructor.

Response Lenses

mrsgrsReplicationSubnetGroup :: Lens' ModifyReplicationSubnetGroupResponse (Maybe ReplicationSubnetGroup)

The modified replication subnet group.

mrsgrsResponseStatus :: Lens' ModifyReplicationSubnetGroupResponse Int

- | The response status code.