Network.AWS.DMS.ModifyReplicationTask

Description

Modifies the specified replication task.

You can't modify the task endpoints. The task must be stopped before you can modify it.

For more information about AWS DMS tasks, see the AWS DMS user guide at Working with Migration Tasks

Creating a Request

modifyReplicationTask

Arguments

:: Text
mrtReplicationTaskARN

-> ModifyReplicationTask

Creates a value of ModifyReplicationTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mrtReplicationTaskSettings - JSON file that contains settings for the task, such as target metadata settings.

mrtReplicationTaskIdentifier - The replication task identifier. Constraints: * Must contain from 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

mrtTableMappings - When using the AWS CLI or boto3, provide the path of the JSON file that contains the table mappings. Precede the path with "file:/". When working with the DMS API, provide the JSON as the parameter value. For example, --table-mappings file:/mappingfile.json

mrtMigrationType - The migration type. Valid values: full-load | cdc | full-load-and-cdc

mrtCdcStartTime - The start time for the Change Data Capture (CDC) operation.

mrtReplicationTaskARN - The Amazon Resource Name (ARN) of the replication task.

data ModifyReplicationTask

See: modifyReplicationTask smart constructor.

Request Lenses

mrtReplicationTaskSettings :: Lens' ModifyReplicationTask (Maybe Text)

JSON file that contains settings for the task, such as target metadata settings.

mrtReplicationTaskIdentifier :: Lens' ModifyReplicationTask (Maybe Text)

The replication task identifier. Constraints: * Must contain from 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

mrtTableMappings :: Lens' ModifyReplicationTask (Maybe Text)

When using the AWS CLI or boto3, provide the path of the JSON file that contains the table mappings. Precede the path with "file:/". When working with the DMS API, provide the JSON as the parameter value. For example, --table-mappings file:/mappingfile.json

mrtMigrationType :: Lens' ModifyReplicationTask (Maybe MigrationTypeValue)

The migration type. Valid values: full-load | cdc | full-load-and-cdc

mrtCdcStartTime :: Lens' ModifyReplicationTask (Maybe UTCTime)

The start time for the Change Data Capture (CDC) operation.

mrtReplicationTaskARN :: Lens' ModifyReplicationTask Text

The Amazon Resource Name (ARN) of the replication task.

Destructuring the Response

modifyReplicationTaskResponse

Arguments

:: Int
mrtrsResponseStatus

-> ModifyReplicationTaskResponse

Creates a value of ModifyReplicationTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mrtrsReplicationTask - The replication task that was modified.

mrtrsResponseStatus - -- | The response status code.

data ModifyReplicationTaskResponse

See: modifyReplicationTaskResponse smart constructor.

Response Lenses

mrtrsReplicationTask :: Lens' ModifyReplicationTaskResponse (Maybe ReplicationTask)

The replication task that was modified.

mrtrsResponseStatus :: Lens' ModifyReplicationTaskResponse Int

- | The response status code.