Network.AWS.DMS.RefreshSchemas

Description

Populates the schema for the specified endpoint. This is an asynchronous operation and can take several minutes. You can check the status of this operation by calling the DescribeRefreshSchemasStatus operation.

Creating a Request

refreshSchemas

Arguments

:: Text
rsEndpointARN

-> Text
rsReplicationInstanceARN

-> RefreshSchemas

Creates a value of RefreshSchemas with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsEndpointARN - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

rsReplicationInstanceARN - The Amazon Resource Name (ARN) of the replication instance.

data RefreshSchemas

See: refreshSchemas smart constructor.

Request Lenses

rsEndpointARN :: Lens' RefreshSchemas Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

rsReplicationInstanceARN :: Lens' RefreshSchemas Text

The Amazon Resource Name (ARN) of the replication instance.

Destructuring the Response

refreshSchemasResponse

Arguments

:: Int
rsrsResponseStatus

-> RefreshSchemasResponse

Creates a value of RefreshSchemasResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsrsRefreshSchemasStatus - The status of the refreshed schema.

rsrsResponseStatus - -- | The response status code.

data RefreshSchemasResponse

See: refreshSchemasResponse smart constructor.

Response Lenses

rsrsRefreshSchemasStatus :: Lens' RefreshSchemasResponse (Maybe RefreshSchemasStatus)

The status of the refreshed schema.

rsrsResponseStatus :: Lens' RefreshSchemasResponse Int

- | The response status code.