Network.AWS.DMS.TestConnection

Description

Tests the connection between the replication instance and the endpoint.

Creating a Request

testConnection

Arguments

:: Text
tcReplicationInstanceARN

-> Text
tcEndpointARN

-> TestConnection

Creates a value of TestConnection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tcReplicationInstanceARN - The Amazon Resource Name (ARN) of the replication instance.

tcEndpointARN - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

data TestConnection

See: testConnection smart constructor.

Request Lenses

tcReplicationInstanceARN :: Lens' TestConnection Text

The Amazon Resource Name (ARN) of the replication instance.

tcEndpointARN :: Lens' TestConnection Text

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

Destructuring the Response

testConnectionResponse

Arguments

:: Int
tcrsResponseStatus

-> TestConnectionResponse

Creates a value of TestConnectionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tcrsConnection - The connection tested.

tcrsResponseStatus - -- | The response status code.

data TestConnectionResponse

See: testConnectionResponse smart constructor.

Response Lenses

tcrsConnection :: Lens' TestConnectionResponse (Maybe Connection)

The connection tested.

tcrsResponseStatus :: Lens' TestConnectionResponse Int

- | The response status code.