Network.AWS.DMS.Types

Description

Service Configuration

dms :: Service

API version 2016-01-01 of the Amazon Database Migration Service SDK configuration.

Errors

_InvalidSubnet :: AsError a => Getting (First ServiceError) a ServiceError

The subnet provided is invalid.

_KMSKeyNotAccessibleFault :: AsError a => Getting (First ServiceError) a ServiceError

AWS DMS cannot access the KMS key.

_ReplicationSubnetGroupDoesNotCoverEnoughAZs :: AsError a => Getting (First ServiceError) a ServiceError

The replication subnet group does not cover enough Availability Zones (AZs). Edit the replication subnet group and add more AZs.

_InvalidResourceStateFault :: AsError a => Getting (First ServiceError) a ServiceError

The resource is in a state that prevents it from being used for database migration.

_InvalidCertificateFault :: AsError a => Getting (First ServiceError) a ServiceError

The certificate was not valid.

_SNSNoAuthorizationFault :: AsError a => Getting (First ServiceError) a ServiceError

You are not authorized for the SNS subscription.

_ResourceAlreadyExistsFault :: AsError a => Getting (First ServiceError) a ServiceError

The resource you are attempting to create already exists.

_InsufficientResourceCapacityFault :: AsError a => Getting (First ServiceError) a ServiceError

There are not enough resources allocated to the database migration.

_SNSInvalidTopicFault :: AsError a => Getting (First ServiceError) a ServiceError

The SNS topic is invalid.

_ResourceQuotaExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

The quota for this resource quota has been exceeded.

_UpgradeDependencyFailureFault :: AsError a => Getting (First ServiceError) a ServiceError

An upgrade dependency is preventing the database migration.

_ResourceNotFoundFault :: AsError a => Getting (First ServiceError) a ServiceError

The resource could not be found.

_StorageQuotaExceededFault :: AsError a => Getting (First ServiceError) a ServiceError

The storage quota has been exceeded.

_AccessDeniedFault :: AsError a => Getting (First ServiceError) a ServiceError

AWS DMS was denied access to the endpoint.

_SubnetAlreadyInUse :: AsError a => Getting (First ServiceError) a ServiceError

The specified subnet is already in use.

AuthMechanismValue

data AuthMechanismValue

Constructors

Default

MongodbCr

ScramSha1

AuthTypeValue

data AuthTypeValue

Constructors

NO

Password

CompressionTypeValue

data CompressionTypeValue

Constructors

CTVGzip

CTVNone

DmsSSLModeValue

data DmsSSLModeValue

Constructors

DSMVNone

DSMVRequire

DSMVVerifyCa

DSMVVerifyFull

MigrationTypeValue

data MigrationTypeValue

Constructors

Cdc

FullLoad

FullLoadAndCdc

NestingLevelValue

data NestingLevelValue

Constructors

None

One

RefreshSchemasStatusTypeValue

data RefreshSchemasStatusTypeValue

Constructors

Failed

Refreshing

Successful

ReplicationEndpointTypeValue

data ReplicationEndpointTypeValue

Constructors

Source

Target

SourceType

data SourceType

Constructors

ReplicationInstance

StartReplicationTaskTypeValue

data StartReplicationTaskTypeValue

Constructors

ReloadTarget

ResumeProcessing

StartReplication

AccountQuota

data AccountQuota

Describes a quota for an AWS account, for example, the number of replication instances allowed.

See: accountQuota smart constructor.

accountQuota :: AccountQuota

Creates a value of AccountQuota with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aqMax - The maximum allowed value for the quota.

aqUsed - The amount currently used toward the quota maximum.

aqAccountQuotaName - The name of the AWS DMS quota for this AWS account.

aqMax :: Lens' AccountQuota (Maybe Integer)

The maximum allowed value for the quota.

aqUsed :: Lens' AccountQuota (Maybe Integer)

The amount currently used toward the quota maximum.

aqAccountQuotaName :: Lens' AccountQuota (Maybe Text)

The name of the AWS DMS quota for this AWS account.

AvailabilityZone

data AvailabilityZone

See: availabilityZone smart constructor.

availabilityZone :: AvailabilityZone

Creates a value of AvailabilityZone with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

azName - The name of the availability zone.

azName :: Lens' AvailabilityZone (Maybe Text)

The name of the availability zone.

Certificate

data Certificate

The SSL certificate that can be used to encrypt connections between the endpoints and the replication instance.

See: certificate smart constructor.

certificate :: Certificate

Creates a value of Certificate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cCertificateOwner - The owner of the certificate.

cSigningAlgorithm - The signing algorithm for the certificate.

cValidFromDate - The beginning date that the certificate is valid.

cCertificatePem - The contents of the .pem X.509 certificate file for the certificate.

cCertificateARN - The Amazon Resource Name (ARN) for the certificate.

cCertificateCreationDate - The date that the certificate was created.

cCertificateIdentifier - The customer-assigned name of the certificate. Valid characters are A-z and 0-9.

cCertificateWallet - The location of the imported Oracle Wallet certificate for use with SSL.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

cKeyLength - The key length of the cryptographic algorithm being used.

cValidToDate - The final date that the certificate is valid.

cCertificateOwner :: Lens' Certificate (Maybe Text)

The owner of the certificate.

cSigningAlgorithm :: Lens' Certificate (Maybe Text)

The signing algorithm for the certificate.

cValidFromDate :: Lens' Certificate (Maybe UTCTime)

The beginning date that the certificate is valid.

cCertificatePem :: Lens' Certificate (Maybe Text)

The contents of the .pem X.509 certificate file for the certificate.

cCertificateARN :: Lens' Certificate (Maybe Text)

The Amazon Resource Name (ARN) for the certificate.

cCertificateCreationDate :: Lens' Certificate (Maybe UTCTime)

The date that the certificate was created.

cCertificateIdentifier :: Lens' Certificate (Maybe Text)

The customer-assigned name of the certificate. Valid characters are A-z and 0-9.

cCertificateWallet :: Lens' Certificate (Maybe ByteString)

The location of the imported Oracle Wallet certificate for use with SSL.-- Note: This Lens automatically encodes and decodes Base64 data. The underlying isomorphism will encode to Base64 representation during serialisation, and decode from Base64 representation during deserialisation. This Lens accepts and returns only raw unencoded data.

cKeyLength :: Lens' Certificate (Maybe Int)

The key length of the cryptographic algorithm being used.

cValidToDate :: Lens' Certificate (Maybe UTCTime)

The final date that the certificate is valid.

Connection

data Connection

See: connection smart constructor.

connection :: Connection

Creates a value of Connection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cStatus - The connection status.

cReplicationInstanceARN - The Amazon Resource Name (ARN) of the replication instance.

cEndpointIdentifier - The identifier of the endpoint. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.

cReplicationInstanceIdentifier - The replication instance identifier. This parameter is stored as a lowercase string.

cEndpointARN - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

cLastFailureMessage - The error message when the connection last failed.

cStatus :: Lens' Connection (Maybe Text)

The connection status.

cReplicationInstanceARN :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) of the replication instance.

cEndpointIdentifier :: Lens' Connection (Maybe Text)

The identifier of the endpoint. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.

cReplicationInstanceIdentifier :: Lens' Connection (Maybe Text)

The replication instance identifier. This parameter is stored as a lowercase string.

cEndpointARN :: Lens' Connection (Maybe Text)

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

cLastFailureMessage :: Lens' Connection (Maybe Text)

The error message when the connection last failed.

DynamoDBSettings

data DynamoDBSettings

See: dynamoDBSettings smart constructor.

dynamoDBSettings

Arguments

:: Text
ddsServiceAccessRoleARN

-> DynamoDBSettings

Creates a value of DynamoDBSettings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddsServiceAccessRoleARN - The Amazon Resource Name (ARN) used by the service access IAM role.

ddsServiceAccessRoleARN :: Lens' DynamoDBSettings Text

The Amazon Resource Name (ARN) used by the service access IAM role.

Endpoint

data Endpoint

See: endpoint smart constructor.

endpoint :: Endpoint

Creates a value of Endpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eStatus - The status of the endpoint.

eServerName - The name of the server at the endpoint.

eCertificateARN - The Amazon Resource Name (ARN) used for SSL connection to the endpoint.

eExtraConnectionAttributes - Additional connection attributes used to connect to the endpoint.

eEndpointType - The type of endpoint.

eUsername - The user name used to connect to the endpoint.

eEngineName - The database engine name. Valid values, depending on the EndPointType, include MYSQL, ORACLE, POSTGRES, MARIADB, AURORA, REDSHIFT, S3, SYBASE, DYNAMODB, MONGODB, and SQLSERVER.

eKMSKeyId - The KMS key identifier that will be used to encrypt the connection parameters. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

eMongoDBSettings - The settings for the MongoDB source endpoint. For more information, see the MongoDbSettings structure.

eSSLMode - The SSL mode used to connect to the endpoint. SSL mode can be one of four values: none, require, verify-ca, verify-full. The default value is none.

eDatabaseName - The name of the database at the endpoint.

eS3Settings - The settings for the S3 target endpoint. For more information, see the S3Settings structure.

eEndpointIdentifier - The database endpoint identifier. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.

eExternalId - Value returned by a call to CreateEndpoint that can be used for cross-account validation. Use it on a subsequent call to CreateEndpoint to create the endpoint with a cross-account.

eDynamoDBSettings - The settings for the target DynamoDB database. For more information, see the DynamoDBSettings structure.

eEndpointARN - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

ePort - The port value used to access the endpoint.

eStatus :: Lens' Endpoint (Maybe Text)

The status of the endpoint.

eServerName :: Lens' Endpoint (Maybe Text)

The name of the server at the endpoint.

eCertificateARN :: Lens' Endpoint (Maybe Text)

The Amazon Resource Name (ARN) used for SSL connection to the endpoint.

eExtraConnectionAttributes :: Lens' Endpoint (Maybe Text)

Additional connection attributes used to connect to the endpoint.

eEndpointType :: Lens' Endpoint (Maybe ReplicationEndpointTypeValue)

The type of endpoint.

eUsername :: Lens' Endpoint (Maybe Text)

The user name used to connect to the endpoint.

eEngineName :: Lens' Endpoint (Maybe Text)

The database engine name. Valid values, depending on the EndPointType, include MYSQL, ORACLE, POSTGRES, MARIADB, AURORA, REDSHIFT, S3, SYBASE, DYNAMODB, MONGODB, and SQLSERVER.

eKMSKeyId :: Lens' Endpoint (Maybe Text)

The KMS key identifier that will be used to encrypt the connection parameters. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

eMongoDBSettings :: Lens' Endpoint (Maybe MongoDBSettings)

The settings for the MongoDB source endpoint. For more information, see the MongoDbSettings structure.

eSSLMode :: Lens' Endpoint (Maybe DmsSSLModeValue)

The SSL mode used to connect to the endpoint. SSL mode can be one of four values: none, require, verify-ca, verify-full. The default value is none.

eDatabaseName :: Lens' Endpoint (Maybe Text)

The name of the database at the endpoint.

eS3Settings :: Lens' Endpoint (Maybe S3Settings)

The settings for the S3 target endpoint. For more information, see the S3Settings structure.

eEndpointIdentifier :: Lens' Endpoint (Maybe Text)

The database endpoint identifier. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.

eExternalId :: Lens' Endpoint (Maybe Text)

Value returned by a call to CreateEndpoint that can be used for cross-account validation. Use it on a subsequent call to CreateEndpoint to create the endpoint with a cross-account.

eDynamoDBSettings :: Lens' Endpoint (Maybe DynamoDBSettings)

The settings for the target DynamoDB database. For more information, see the DynamoDBSettings structure.

eEndpointARN :: Lens' Endpoint (Maybe Text)

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

ePort :: Lens' Endpoint (Maybe Int)

The port value used to access the endpoint.

Event

data Event

See: event smart constructor.

event :: Event

Creates a value of Event with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

eSourceType - The type of AWS DMS resource that generates events. Valid values: replication-instance | endpoint | migration-task

eSourceIdentifier - The identifier of the event source. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it cannot end with a hyphen or contain two consecutive hyphens. Constraints:replication instance, endpoint, migration task

eDate - The date of the event.

eEventCategories - The event categories available for the specified source type.

eMessage - The event message.

eSourceType :: Lens' Event (Maybe SourceType)

The type of AWS DMS resource that generates events. Valid values: replication-instance | endpoint | migration-task

eSourceIdentifier :: Lens' Event (Maybe Text)

The identifier of the event source. An identifier must begin with a letter and must contain only ASCII letters, digits, and hyphens; it cannot end with a hyphen or contain two consecutive hyphens. Constraints:replication instance, endpoint, migration task

eDate :: Lens' Event (Maybe UTCTime)

The date of the event.

eEventCategories :: Lens' Event [Text]

The event categories available for the specified source type.

eMessage :: Lens' Event (Maybe Text)

The event message.

EventCategoryGroup

data EventCategoryGroup

See: eventCategoryGroup smart constructor.

eventCategoryGroup :: EventCategoryGroup

Creates a value of EventCategoryGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ecgSourceType - The type of AWS DMS resource that generates events. Valid values: replication-instance | replication-server | security-group | migration-task

ecgEventCategories - A list of event categories for a SourceType that you want to subscribe to.

ecgSourceType :: Lens' EventCategoryGroup (Maybe Text)

The type of AWS DMS resource that generates events. Valid values: replication-instance | replication-server | security-group | migration-task

ecgEventCategories :: Lens' EventCategoryGroup [Text]

A list of event categories for a SourceType that you want to subscribe to.

EventSubscription

data EventSubscription

See: eventSubscription smart constructor.

eventSubscription :: EventSubscription

Creates a value of EventSubscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esStatus - The status of the AWS DMS event notification subscription. Constraints: Can be one of the following: creating | modifying | deleting | active | no-permission | topic-not-exist The status "no-permission" indicates that AWS DMS no longer has permission to post to the SNS topic. The status "topic-not-exist" indicates that the topic was deleted after the subscription was created.

esCustomerAWSId - The AWS customer account associated with the AWS DMS event notification subscription.

esCustSubscriptionId - The AWS DMS event notification subscription Id.

esSNSTopicARN - The topic ARN of the AWS DMS event notification subscription.

esEnabled - Boolean value that indicates if the event subscription is enabled.

esSourceType - The type of AWS DMS resource that generates events. Valid values: replication-instance | replication-server | security-group | migration-task

esSubscriptionCreationTime - The time the RDS event notification subscription was created.

esEventCategoriesList - A lists of event categories.

esSourceIdsList - A list of source Ids for the event subscription.

esStatus :: Lens' EventSubscription (Maybe Text)

The status of the AWS DMS event notification subscription. Constraints: Can be one of the following: creating | modifying | deleting | active | no-permission | topic-not-exist The status "no-permission" indicates that AWS DMS no longer has permission to post to the SNS topic. The status "topic-not-exist" indicates that the topic was deleted after the subscription was created.

esCustomerAWSId :: Lens' EventSubscription (Maybe Text)

The AWS customer account associated with the AWS DMS event notification subscription.

esCustSubscriptionId :: Lens' EventSubscription (Maybe Text)

The AWS DMS event notification subscription Id.

esSNSTopicARN :: Lens' EventSubscription (Maybe Text)

The topic ARN of the AWS DMS event notification subscription.

esEnabled :: Lens' EventSubscription (Maybe Bool)

Boolean value that indicates if the event subscription is enabled.

esSourceType :: Lens' EventSubscription (Maybe Text)

The type of AWS DMS resource that generates events. Valid values: replication-instance | replication-server | security-group | migration-task

esSubscriptionCreationTime :: Lens' EventSubscription (Maybe Text)

The time the RDS event notification subscription was created.

esEventCategoriesList :: Lens' EventSubscription [Text]

A lists of event categories.

esSourceIdsList :: Lens' EventSubscription [Text]

A list of source Ids for the event subscription.

Filter

data Filter

See: filter' smart constructor.

filter'

Arguments

:: Text
fName

-> Filter

Creates a value of Filter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fName - The name of the filter.

fValues - The filter value.

fName :: Lens' Filter Text

The name of the filter.

fValues :: Lens' Filter [Text]

The filter value.

MongoDBSettings

data MongoDBSettings

See: mongoDBSettings smart constructor.

mongoDBSettings :: MongoDBSettings

Creates a value of MongoDBSettings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mdsServerName - The name of the server on the MongoDB source endpoint.

mdsAuthMechanism - The authentication mechanism you use to access the MongoDB source endpoint. Valid values: DEFAULT, MONGODB_CR, SCRAM_SHA_1 DEFAULT – For MongoDB version 2.x, use MONGODB_CR. For MongoDB version 3.x, use SCRAM_SHA_1. This attribute is not used when authType=No.

mdsUsername - The user name you use to access the MongoDB source endpoint.

mdsPassword - The password for the user account you use to access the MongoDB source endpoint.

mdsNestingLevel - Specifies either document or table mode. Valid values: NONE, ONE Default value is NONE. Specify NONE to use document mode. Specify ONE to use table mode.

mdsDatabaseName - The database name on the MongoDB source endpoint.

mdsDocsToInvestigate - Indicates the number of documents to preview to determine the document organization. Use this attribute when NestingLevel is set to ONE. Must be a positive value greater than 0. Default value is 1000.

mdsAuthSource - The MongoDB database name. This attribute is not used when authType=NO . The default is admin.

mdsExtractDocId - Specifies the document ID. Use this attribute when NestingLevel is set to NONE. Default value is false.

mdsAuthType - The authentication type you use to access the MongoDB source endpoint. Valid values: NO, PASSWORD When NO is selected, user name and password parameters are not used and can be empty.

mdsPort - The port value for the MongoDB source endpoint.

mdsServerName :: Lens' MongoDBSettings (Maybe Text)

The name of the server on the MongoDB source endpoint.

mdsAuthMechanism :: Lens' MongoDBSettings (Maybe AuthMechanismValue)

The authentication mechanism you use to access the MongoDB source endpoint. Valid values: DEFAULT, MONGODB_CR, SCRAM_SHA_1 DEFAULT – For MongoDB version 2.x, use MONGODB_CR. For MongoDB version 3.x, use SCRAM_SHA_1. This attribute is not used when authType=No.

mdsUsername :: Lens' MongoDBSettings (Maybe Text)

The user name you use to access the MongoDB source endpoint.

mdsPassword :: Lens' MongoDBSettings (Maybe Text)

The password for the user account you use to access the MongoDB source endpoint.

mdsNestingLevel :: Lens' MongoDBSettings (Maybe NestingLevelValue)

Specifies either document or table mode. Valid values: NONE, ONE Default value is NONE. Specify NONE to use document mode. Specify ONE to use table mode.

mdsDatabaseName :: Lens' MongoDBSettings (Maybe Text)

The database name on the MongoDB source endpoint.

mdsDocsToInvestigate :: Lens' MongoDBSettings (Maybe Text)

Indicates the number of documents to preview to determine the document organization. Use this attribute when NestingLevel is set to ONE. Must be a positive value greater than 0. Default value is 1000.

mdsAuthSource :: Lens' MongoDBSettings (Maybe Text)

The MongoDB database name. This attribute is not used when authType=NO . The default is admin.

mdsExtractDocId :: Lens' MongoDBSettings (Maybe Text)

Specifies the document ID. Use this attribute when NestingLevel is set to NONE. Default value is false.

mdsAuthType :: Lens' MongoDBSettings (Maybe AuthTypeValue)

The authentication type you use to access the MongoDB source endpoint. Valid values: NO, PASSWORD When NO is selected, user name and password parameters are not used and can be empty.

mdsPort :: Lens' MongoDBSettings (Maybe Int)

The port value for the MongoDB source endpoint.

OrderableReplicationInstance

data OrderableReplicationInstance

See: orderableReplicationInstance smart constructor.

orderableReplicationInstance :: OrderableReplicationInstance

Creates a value of OrderableReplicationInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

oriEngineVersion - The version of the replication engine.

oriMinAllocatedStorage - The minimum amount of storage (in gigabytes) that can be allocated for the replication instance.

oriIncludedAllocatedStorage - The amount of storage (in gigabytes) that is allocated for the replication instance.

oriMaxAllocatedStorage - The minimum amount of storage (in gigabytes) that can be allocated for the replication instance.

oriReplicationInstanceClass - The compute and memory capacity of the replication instance. Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

oriDefaultAllocatedStorage - The default amount of storage (in gigabytes) that is allocated for the replication instance.

oriStorageType - The type of storage used by the replication instance.

oriEngineVersion :: Lens' OrderableReplicationInstance (Maybe Text)

The version of the replication engine.

oriMinAllocatedStorage :: Lens' OrderableReplicationInstance (Maybe Int)

The minimum amount of storage (in gigabytes) that can be allocated for the replication instance.

oriIncludedAllocatedStorage :: Lens' OrderableReplicationInstance (Maybe Int)

The amount of storage (in gigabytes) that is allocated for the replication instance.

oriMaxAllocatedStorage :: Lens' OrderableReplicationInstance (Maybe Int)

The minimum amount of storage (in gigabytes) that can be allocated for the replication instance.

oriReplicationInstanceClass :: Lens' OrderableReplicationInstance (Maybe Text)

The compute and memory capacity of the replication instance. Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

oriDefaultAllocatedStorage :: Lens' OrderableReplicationInstance (Maybe Int)

The default amount of storage (in gigabytes) that is allocated for the replication instance.

oriStorageType :: Lens' OrderableReplicationInstance (Maybe Text)

The type of storage used by the replication instance.

RefreshSchemasStatus

data RefreshSchemasStatus

See: refreshSchemasStatus smart constructor.

refreshSchemasStatus :: RefreshSchemasStatus

Creates a value of RefreshSchemasStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rssStatus - The status of the schema.

rssLastRefreshDate - The date the schema was last refreshed.

rssReplicationInstanceARN - The Amazon Resource Name (ARN) of the replication instance.

rssEndpointARN - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

rssLastFailureMessage - The last failure message for the schema.

rssStatus :: Lens' RefreshSchemasStatus (Maybe RefreshSchemasStatusTypeValue)

The status of the schema.

rssLastRefreshDate :: Lens' RefreshSchemasStatus (Maybe UTCTime)

The date the schema was last refreshed.

rssReplicationInstanceARN :: Lens' RefreshSchemasStatus (Maybe Text)

The Amazon Resource Name (ARN) of the replication instance.

rssEndpointARN :: Lens' RefreshSchemasStatus (Maybe Text)

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

rssLastFailureMessage :: Lens' RefreshSchemasStatus (Maybe Text)

The last failure message for the schema.

ReplicationInstance

data ReplicationInstance

See: replicationInstance smart constructor.

replicationInstance :: ReplicationInstance

Creates a value of ReplicationInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

riEngineVersion - The engine version number of the replication instance.

riPubliclyAccessible - Specifies the accessibility options for the replication instance. A value of true represents an instance with a public IP address. A value of false represents an instance with a private IP address. The default value is true .

riAutoMinorVersionUpgrade - Boolean value indicating if minor version upgrades will be automatically applied to the instance.

riReplicationInstancePublicIPAddresses - The public IP address of the replication instance.

riReplicationSubnetGroup - The subnet group for the replication instance.

riInstanceCreateTime - The time the replication instance was created.

riReplicationInstanceStatus - The status of the replication instance.

riReplicationInstancePrivateIPAddresses - The private IP address of the replication instance.

riPreferredMaintenanceWindow - The maintenance window times for the replication instance.

riReplicationInstancePrivateIPAddress - The private IP address of the replication instance.

riKMSKeyId - The KMS key identifier that is used to encrypt the content on the replication instance. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

riAvailabilityZone - The Availability Zone for the instance.

riVPCSecurityGroups - The VPC security group for the instance.

riMultiAZ - Specifies if the replication instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the Multi-AZ parameter is set to true .

riSecondaryAvailabilityZone - The availability zone of the standby replication instance in a Multi-AZ deployment.

riReplicationInstanceARN - The Amazon Resource Name (ARN) of the replication instance.

riAllocatedStorage - The amount of storage (in gigabytes) that is allocated for the replication instance.

riReplicationInstancePublicIPAddress - The public IP address of the replication instance.

riReplicationInstanceClass - The compute and memory capacity of the replication instance. Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

riReplicationInstanceIdentifier - The replication instance identifier. This parameter is stored as a lowercase string. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. Example: myrepinstance

riPendingModifiedValues - The pending modification values.

riEngineVersion :: Lens' ReplicationInstance (Maybe Text)

The engine version number of the replication instance.

riPubliclyAccessible :: Lens' ReplicationInstance (Maybe Bool)

Specifies the accessibility options for the replication instance. A value of true represents an instance with a public IP address. A value of false represents an instance with a private IP address. The default value is true .

riAutoMinorVersionUpgrade :: Lens' ReplicationInstance (Maybe Bool)

Boolean value indicating if minor version upgrades will be automatically applied to the instance.

riReplicationInstancePublicIPAddresses :: Lens' ReplicationInstance [Text]

The public IP address of the replication instance.

riReplicationSubnetGroup :: Lens' ReplicationInstance (Maybe ReplicationSubnetGroup)

The subnet group for the replication instance.

riInstanceCreateTime :: Lens' ReplicationInstance (Maybe UTCTime)

The time the replication instance was created.

riReplicationInstanceStatus :: Lens' ReplicationInstance (Maybe Text)

The status of the replication instance.

riReplicationInstancePrivateIPAddresses :: Lens' ReplicationInstance [Text]

The private IP address of the replication instance.

riPreferredMaintenanceWindow :: Lens' ReplicationInstance (Maybe Text)

The maintenance window times for the replication instance.

riReplicationInstancePrivateIPAddress :: Lens' ReplicationInstance (Maybe Text)

The private IP address of the replication instance.

riKMSKeyId :: Lens' ReplicationInstance (Maybe Text)

The KMS key identifier that is used to encrypt the content on the replication instance. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

riAvailabilityZone :: Lens' ReplicationInstance (Maybe Text)

The Availability Zone for the instance.

riVPCSecurityGroups :: Lens' ReplicationInstance [VPCSecurityGroupMembership]

The VPC security group for the instance.

riMultiAZ :: Lens' ReplicationInstance (Maybe Bool)

Specifies if the replication instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the Multi-AZ parameter is set to true .

riSecondaryAvailabilityZone :: Lens' ReplicationInstance (Maybe Text)

The availability zone of the standby replication instance in a Multi-AZ deployment.

riReplicationInstanceARN :: Lens' ReplicationInstance (Maybe Text)

The Amazon Resource Name (ARN) of the replication instance.

riAllocatedStorage :: Lens' ReplicationInstance (Maybe Int)

The amount of storage (in gigabytes) that is allocated for the replication instance.

riReplicationInstancePublicIPAddress :: Lens' ReplicationInstance (Maybe Text)

The public IP address of the replication instance.

riReplicationInstanceClass :: Lens' ReplicationInstance (Maybe Text)

The compute and memory capacity of the replication instance. Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

riReplicationInstanceIdentifier :: Lens' ReplicationInstance (Maybe Text)

The replication instance identifier. This parameter is stored as a lowercase string. Constraints: * Must contain from 1 to 63 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens. Example: myrepinstance

riPendingModifiedValues :: Lens' ReplicationInstance (Maybe ReplicationPendingModifiedValues)

The pending modification values.

ReplicationPendingModifiedValues

data ReplicationPendingModifiedValues

See: replicationPendingModifiedValues smart constructor.

replicationPendingModifiedValues :: ReplicationPendingModifiedValues

Creates a value of ReplicationPendingModifiedValues with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rpmvEngineVersion - The engine version number of the replication instance.

rpmvMultiAZ - Specifies if the replication instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the Multi-AZ parameter is set to true .

rpmvAllocatedStorage - The amount of storage (in gigabytes) that is allocated for the replication instance.

rpmvReplicationInstanceClass - The compute and memory capacity of the replication instance. Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

rpmvEngineVersion :: Lens' ReplicationPendingModifiedValues (Maybe Text)

The engine version number of the replication instance.

rpmvMultiAZ :: Lens' ReplicationPendingModifiedValues (Maybe Bool)

Specifies if the replication instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the Multi-AZ parameter is set to true .

rpmvAllocatedStorage :: Lens' ReplicationPendingModifiedValues (Maybe Int)

The amount of storage (in gigabytes) that is allocated for the replication instance.

rpmvReplicationInstanceClass :: Lens' ReplicationPendingModifiedValues (Maybe Text)

The compute and memory capacity of the replication instance. Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

ReplicationSubnetGroup

data ReplicationSubnetGroup

See: replicationSubnetGroup smart constructor.

replicationSubnetGroup :: ReplicationSubnetGroup

Creates a value of ReplicationSubnetGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsgVPCId - The ID of the VPC.

rsgSubnets - The subnets that are in the subnet group.

rsgReplicationSubnetGroupIdentifier - The identifier of the replication instance subnet group.

rsgSubnetGroupStatus - The status of the subnet group.

rsgReplicationSubnetGroupDescription - The description of the replication subnet group.

rsgVPCId :: Lens' ReplicationSubnetGroup (Maybe Text)

The ID of the VPC.

rsgSubnets :: Lens' ReplicationSubnetGroup [Subnet]

The subnets that are in the subnet group.

rsgReplicationSubnetGroupIdentifier :: Lens' ReplicationSubnetGroup (Maybe Text)

The identifier of the replication instance subnet group.

rsgSubnetGroupStatus :: Lens' ReplicationSubnetGroup (Maybe Text)

The status of the subnet group.

rsgReplicationSubnetGroupDescription :: Lens' ReplicationSubnetGroup (Maybe Text)

The description of the replication subnet group.

ReplicationTask

data ReplicationTask

See: replicationTask smart constructor.

replicationTask :: ReplicationTask

Creates a value of ReplicationTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rReplicationTaskSettings - The settings for the replication task.

rStatus - The status of the replication task.

rStopReason - The reason the replication task was stopped.

rTargetEndpointARN - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

rReplicationTaskIdentifier - The replication task identifier. Constraints: * Must contain from 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

rReplicationTaskStartDate - The date the replication task is scheduled to start.

rSourceEndpointARN - The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

rTableMappings - Table mappings specified in the task.

rReplicationTaskCreationDate - The date the replication task was created.

rMigrationType - The type of migration.

rReplicationTaskARN - The Amazon Resource Name (ARN) of the replication task.

rReplicationTaskStats - The statistics for the task, including elapsed time, tables loaded, and table errors.

rReplicationInstanceARN - The Amazon Resource Name (ARN) of the replication instance.

rLastFailureMessage - The last error (failure) message generated for the replication instance.

rReplicationTaskSettings :: Lens' ReplicationTask (Maybe Text)

The settings for the replication task.

rStatus :: Lens' ReplicationTask (Maybe Text)

The status of the replication task.

rStopReason :: Lens' ReplicationTask (Maybe Text)

The reason the replication task was stopped.

rTargetEndpointARN :: Lens' ReplicationTask (Maybe Text)

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

rReplicationTaskIdentifier :: Lens' ReplicationTask (Maybe Text)

The replication task identifier. Constraints: * Must contain from 1 to 255 alphanumeric characters or hyphens. * First character must be a letter. * Cannot end with a hyphen or contain two consecutive hyphens.

rReplicationTaskStartDate :: Lens' ReplicationTask (Maybe UTCTime)

The date the replication task is scheduled to start.

rSourceEndpointARN :: Lens' ReplicationTask (Maybe Text)

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

rTableMappings :: Lens' ReplicationTask (Maybe Text)

Table mappings specified in the task.

rReplicationTaskCreationDate :: Lens' ReplicationTask (Maybe UTCTime)

The date the replication task was created.

rMigrationType :: Lens' ReplicationTask (Maybe MigrationTypeValue)

The type of migration.

rReplicationTaskARN :: Lens' ReplicationTask (Maybe Text)

The Amazon Resource Name (ARN) of the replication task.

rReplicationTaskStats :: Lens' ReplicationTask (Maybe ReplicationTaskStats)

The statistics for the task, including elapsed time, tables loaded, and table errors.

rReplicationInstanceARN :: Lens' ReplicationTask (Maybe Text)

The Amazon Resource Name (ARN) of the replication instance.

rLastFailureMessage :: Lens' ReplicationTask (Maybe Text)

The last error (failure) message generated for the replication instance.

ReplicationTaskStats

data ReplicationTaskStats

See: replicationTaskStats smart constructor.

replicationTaskStats :: ReplicationTaskStats

Creates a value of ReplicationTaskStats with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtsFullLoadProgressPercent - The percent complete for the full load migration task.

rtsElapsedTimeMillis - The elapsed time of the task, in milliseconds.

rtsTablesErrored - The number of errors that have occurred during this task.

rtsTablesLoaded - The number of tables loaded for this task.

rtsTablesQueued - The number of tables queued for this task.

rtsTablesLoading - The number of tables currently loading for this task.

rtsFullLoadProgressPercent :: Lens' ReplicationTaskStats (Maybe Int)

The percent complete for the full load migration task.

rtsElapsedTimeMillis :: Lens' ReplicationTaskStats (Maybe Integer)

The elapsed time of the task, in milliseconds.

rtsTablesErrored :: Lens' ReplicationTaskStats (Maybe Int)

The number of errors that have occurred during this task.

rtsTablesLoaded :: Lens' ReplicationTaskStats (Maybe Int)

The number of tables loaded for this task.

rtsTablesQueued :: Lens' ReplicationTaskStats (Maybe Int)

The number of tables queued for this task.

rtsTablesLoading :: Lens' ReplicationTaskStats (Maybe Int)

The number of tables currently loading for this task.

S3Settings

data S3Settings

See: s3Settings smart constructor.

s3Settings :: S3Settings

Creates a value of S3Settings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ssCSVDelimiter - The delimiter used to separate columns in the source files. The default is a comma.

ssServiceAccessRoleARN - The Amazon Resource Name (ARN) used by the service access IAM role.

ssBucketFolder - An optional parameter to set a folder name in the S3 bucket. If provided, tables are created in the path bucketFolderschema_nametable_name. If this parameter is not specified, then the path used is schema_nametable_name/.

ssExternalTableDefinition -

ssBucketName - The name of the S3 bucket.

ssCSVRowDelimiter - The delimiter used to separate rows in the source files. The default is a carriage return (n).

ssCompressionType - An optional parameter to use GZIP to compress the target files. Set to GZIP to compress the target files. Set to NONE (the default) or do not use to leave the files uncompressed.

ssCSVDelimiter :: Lens' S3Settings (Maybe Text)

The delimiter used to separate columns in the source files. The default is a comma.

ssServiceAccessRoleARN :: Lens' S3Settings (Maybe Text)

The Amazon Resource Name (ARN) used by the service access IAM role.

ssBucketFolder :: Lens' S3Settings (Maybe Text)

An optional parameter to set a folder name in the S3 bucket. If provided, tables are created in the path bucketFolderschema_nametable_name. If this parameter is not specified, then the path used is schema_nametable_name/.

ssExternalTableDefinition :: Lens' S3Settings (Maybe Text)

ssBucketName :: Lens' S3Settings (Maybe Text)

The name of the S3 bucket.

ssCSVRowDelimiter :: Lens' S3Settings (Maybe Text)

The delimiter used to separate rows in the source files. The default is a carriage return (n).

ssCompressionType :: Lens' S3Settings (Maybe CompressionTypeValue)

An optional parameter to use GZIP to compress the target files. Set to GZIP to compress the target files. Set to NONE (the default) or do not use to leave the files uncompressed.

Subnet

data Subnet

See: subnet smart constructor.

subnet :: Subnet

Creates a value of Subnet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sSubnetStatus - The status of the subnet.

sSubnetIdentifier - The subnet identifier.

sSubnetAvailabilityZone - The Availability Zone of the subnet.

sSubnetStatus :: Lens' Subnet (Maybe Text)

The status of the subnet.

sSubnetIdentifier :: Lens' Subnet (Maybe Text)

The subnet identifier.

sSubnetAvailabilityZone :: Lens' Subnet (Maybe AvailabilityZone)

The Availability Zone of the subnet.

SupportedEndpointType

data SupportedEndpointType

See: supportedEndpointType smart constructor.

supportedEndpointType :: SupportedEndpointType

Creates a value of SupportedEndpointType with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

setEndpointType - The type of endpoint.

setEngineName - The database engine name. Valid values, depending on the EndPointType, include MYSQL, ORACLE, POSTGRES, MARIADB, AURORA, REDSHIFT, S3, SYBASE, DYNAMODB, MONGODB, and SQLSERVER.

setSupportsCDC - Indicates if Change Data Capture (CDC) is supported.

setEndpointType :: Lens' SupportedEndpointType (Maybe ReplicationEndpointTypeValue)

The type of endpoint.

setEngineName :: Lens' SupportedEndpointType (Maybe Text)

The database engine name. Valid values, depending on the EndPointType, include MYSQL, ORACLE, POSTGRES, MARIADB, AURORA, REDSHIFT, S3, SYBASE, DYNAMODB, MONGODB, and SQLSERVER.

setSupportsCDC :: Lens' SupportedEndpointType (Maybe Bool)

Indicates if Change Data Capture (CDC) is supported.

TableStatistics

data TableStatistics

See: tableStatistics smart constructor.

tableStatistics :: TableStatistics

Creates a value of TableStatistics with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tsFullLoadRows - The number of rows added during the Full Load operation.

tsInserts - The number of insert actions performed on a table.

tsFullLoadCondtnlChkFailedRows - The number of rows that failed conditional checks during the Full Load operation (valid only for DynamoDB as a target migrations).

tsSchemaName - The schema name.

tsTableState - The state of the table.

tsFullLoadErrorRows - The number of rows that failed to load during the Full Load operation (valid only for DynamoDB as a target migrations).

tsDdls - The Data Definition Language (DDL) used to build and modify the structure of your tables.

tsDeletes - The number of delete actions performed on a table.

tsUpdates - The number of update actions performed on a table.

tsLastUpdateTime - The last time the table was updated.

tsTableName - The name of the table.

tsFullLoadRows :: Lens' TableStatistics (Maybe Integer)

The number of rows added during the Full Load operation.

tsInserts :: Lens' TableStatistics (Maybe Integer)

The number of insert actions performed on a table.

tsFullLoadCondtnlChkFailedRows :: Lens' TableStatistics (Maybe Integer)

The number of rows that failed conditional checks during the Full Load operation (valid only for DynamoDB as a target migrations).

tsSchemaName :: Lens' TableStatistics (Maybe Text)

The schema name.

tsTableState :: Lens' TableStatistics (Maybe Text)

The state of the table.

tsFullLoadErrorRows :: Lens' TableStatistics (Maybe Integer)

The number of rows that failed to load during the Full Load operation (valid only for DynamoDB as a target migrations).

tsDdls :: Lens' TableStatistics (Maybe Integer)

The Data Definition Language (DDL) used to build and modify the structure of your tables.

tsDeletes :: Lens' TableStatistics (Maybe Integer)

The number of delete actions performed on a table.

tsUpdates :: Lens' TableStatistics (Maybe Integer)

The number of update actions performed on a table.

tsLastUpdateTime :: Lens' TableStatistics (Maybe UTCTime)

The last time the table was updated.

tsTableName :: Lens' TableStatistics (Maybe Text)

The name of the table.

TableToReload

data TableToReload

See: tableToReload smart constructor.

tableToReload :: TableToReload

Creates a value of TableToReload with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ttrSchemaName - The schema name of the table to be reloaded.

ttrTableName - The table name of the table to be reloaded.

ttrSchemaName :: Lens' TableToReload (Maybe Text)

The schema name of the table to be reloaded.

ttrTableName :: Lens' TableToReload (Maybe Text)

The table name of the table to be reloaded.

Tag

data Tag

See: tag smart constructor.

tag :: Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - A value is the optional value of the tag. The string value can be from 1 to 256 Unicode characters in length and cannot be prefixed with "aws:" or "dms:". The string can only contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

tagKey - A key is the required name of the tag. The string value can be from 1 to 128 Unicode characters in length and cannot be prefixed with "aws:" or "dms:". The string can only contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

tagValue :: Lens' Tag (Maybe Text)

A value is the optional value of the tag. The string value can be from 1 to 256 Unicode characters in length and cannot be prefixed with "aws:" or "dms:". The string can only contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

tagKey :: Lens' Tag (Maybe Text)

A key is the required name of the tag. The string value can be from 1 to 128 Unicode characters in length and cannot be prefixed with "aws:" or "dms:". The string can only contain only the set of Unicode letters, digits, white-space, '_', ., /, '=', +, - (Java regex: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

VPCSecurityGroupMembership

data VPCSecurityGroupMembership

See: vpcSecurityGroupMembership smart constructor.

vpcSecurityGroupMembership :: VPCSecurityGroupMembership

Creates a value of VPCSecurityGroupMembership with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vsgmStatus - The status of the VPC security group.

vsgmVPCSecurityGroupId - The VPC security group Id.

vsgmStatus :: Lens' VPCSecurityGroupMembership (Maybe Text)

The status of the VPC security group.

vsgmVPCSecurityGroupId :: Lens' VPCSecurityGroupMembership (Maybe Text)

The VPC security group Id.